Changes to client-side azure resource filtering (#18274)
* slight refactoring

* swapping azure logic to client-side filtering
