+++ src/connectionconfig/azureHelper.ts
@@ -5,9 +5,19 @@
 
 import * as vscode from "vscode";
 import { l10n } from "vscode";
-import { ConnectionDialogWebviewState } from "../sharedInterfaces/connectionDialog";
-import { getErrorMessage } from "../utils/utils";
-import { VSCodeAzureSubscriptionProvider } from "@microsoft/vscode-azext-azureauth";
+import {
+    AzureSqlServerInfo,
+    ConnectionDialogWebviewState,
+} from "../sharedInterfaces/connectionDialog";
+import { getErrorMessage, listAllIterator } from "../utils/utils";
+import {
+    AzureSubscription,
+    VSCodeAzureSubscriptionProvider,
+} from "@microsoft/vscode-azext-azureauth";
+import {
+    GenericResourceExpanded,
+    ResourceManagementClient,
+} from "@azure/arm-resources";
 
 export const azureSubscriptionFilterConfigKey =
     "azureResourceGroups.selectedSubscriptions";
@@ -96,4 +106,77 @@ export async function getQuickPickItems(
         .sort((a, b) => a.label.localeCompare(b.label));
 
     return quickPickItems;
+}
+
+const serverResourceType = "Microsoft.Sql/servers";
+const databaseResourceType = "Microsoft.Sql/servers/databases";
+const elasticPoolsResourceType = "Microsoft.Sql/servers/elasticpools";
+
+export async function fetchServersFromAzure(
+    sub: AzureSubscription,
+): Promise<AzureSqlServerInfo[]> {
+    const result: AzureSqlServerInfo[] = [];
+    const client = new ResourceManagementClient(
+        sub.credential,
+        sub.subscriptionId,
+    );
+
+    // for some subscriptions, supplying a `resourceType eq 'Microsoft.Sql/servers/databases'` filter to list() causes an error:
+    // > invalid filter in query string 'resourceType eq "Microsoft.Sql/servers/databases'"
+    // no idea why, so we're fetching all resources and filtering them ourselves
+
+    const resources = await listAllIterator<GenericResourceExpanded>(
+        client.resources.list(),
+    );
+
+    const servers = resources.filter((r) => r.type === serverResourceType);
+    const databases = resources.filter(
+        (r) =>
+            r.type === databaseResourceType ||
+            r.type === elasticPoolsResourceType,
+    );
+
+    for (const server of servers) {
+        result.push({
+            server: server.name,
+            databases: [],
+            location: server.location,
+            resourceGroup: extractFromResourceId(server.id, "resourceGroups"),
+            subscription: `${sub.name} (${sub.subscriptionId})`,
+        });
+    }
+
+    for (const database of databases) {
+        const serverName = extractFromResourceId(database.id, "servers");
+        const server = result.find((s) => s.server === serverName);
+        if (server) {
+            server.databases.push(
+                database.name.substring(serverName.length + 1),
+            ); // database.name is in the form 'serverName/databaseName', so we need to remove the server name and slash
+        }
+    }
+
+    return result;
+}
+
+function extractFromResourceId(
+    resourceId: string,
+    property: string,
+): string | undefined {
+    if (!property.endsWith("/")) {
+        property += "/";
+    }
+
+    let startIndex = resourceId.indexOf(property);
+
+    if (startIndex === -1) {
+        return undefined;
+    } else {
+        startIndex += property.length;
+    }
+
+    return resourceId.substring(
+        startIndex,
+        resourceId.indexOf("/", startIndex),
+    );
 }
